requirements.txt
duckduckgo_search
gradio
marvin
python-dotenv
